[Android/Feat] 루틴 수정 시 기존 정보 불러오기 완료

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/devices/DeviceListViewModel.kt
@@ -41,7 +41,7 @@ class DeviceListViewModel @Inject constructor(
     @ApplicationContext private val context: Context
 ) : ViewModel() {
 
-    val selectedDeviceId = mutableStateOf<Int?>(null)
+    val selectedDeviceId = mutableStateOf<Long?>(null)
     val showDialog = mutableStateOf(false)
 
 
@@ -181,7 +181,7 @@ class DeviceListViewModel @Inject constructor(
             }
         }
 
-        fun toggleDeviceState(deviceId: Int) {
+        fun toggleDeviceState(deviceId: Long) {
             val currentList = _deviceList.value.toMutableList()
 
             val index = currentList.indexOfFirst { it.deviceId == deviceId }
